+++ vm/state_tree/Cargo.toml
@@ -14,6 +14,9 @@ forest_ipld = "0.1.1"
 ipld_blockstore = "0.1"
 db = { package = "forest_db", version = "0.1" }
 fil_types = { version = "0.2" }
+futures = { version = "0.3" }
+async-std = "1.9"
+forest_car = { path = "../../ipld/car" }
 
 [dev-dependencies]
 num-bigint = { path = "../../utils/bigint", package = "forest_bigint" }
